+++ test/.gitignore
@@ -1,1 +1,2 @@
 a.out
+good-domains.txt
