+++ test/stress.cpp
@@ -171,11 +171,11 @@ int main(int argc, char **argv)
     DNS::Context context(&myloop);
     context.idFactory(&idFactory);
 
-    context.buffersize(24 * 1024 * 1024);      // size of the input buffer (high lowers risk of package loss)
-    context.interval(2.0);                     // number of seconds until the datagram is retried (possibly to next server) (this does not cancel previous requests)
-    context.attempts(4);                       // number of attempts until failure / number of datagrams to send at most
-    context.capacity(10000);                   // max number of simultaneous lookups per dns-context (high increases speed but also risk of package-loss)
-    context.timeout(2.0);                      // time to wait for a response after the _last_ attempt
+    context.buffersize(1024 * 1024); // size of the input buffer (high lowers risk of package loss)
+    context.interval(3.0);           // number of seconds until the datagram is retried (possibly to next server) (this does not cancel previous requests)
+    context.attempts(1);             // number of attempts until failure / number of datagrams to send at most
+    context.capacity(10);            // max number of simultaneous lookups per dns-context (high increases speed but also risk of package-loss)
+    context.timeout(3.0);            // time to wait for a response after the _last_ attempt
 
     // get domain list
     const auto domainlist = readDomainList(argv[1]);
